pyproject.toml
[project]
name = "shinylive_deloy"
version = "2024.7.1"
authors = [
  {name="darrida", email="darrida.py@gmail.com"}
]
description = "Example for non-standard py-shinylive deployments"
readme = "README.md"

dependencies = [
  'shiny',
  'shinylive',
  'gitpython',
  'paramiko',
  'pydantic < 2.0',
]

[project.optional-dependencies]
tests = [
  'fastapi',
  'pytest',
  'loguru',
  'pyfetch_mimic == 2024.06.14',
  'httpx',
  'pytest',
  'shinyswatch',
]

[project.scripts]
shinylive_deploy = "shinylive_deploy.app:main"

[project.urls]
Homepage = "https://github.com/darrida/py-shinylive-deploy-webserver"
Issues = "https://github.com/darrida/py-shinylive-deploy-webserver/issues"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
